requirements.txt
# requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.7.1
flask-limiter==3.8.0
psycopg2-binary==2.9.9
redis==5.0.4
requests==2.32.3
user-agents==2.2.0
qrcode[pil]==8.0
python-dotenv==1.0.1
gunicorn==23.0.0
